Update `clear` button text
The text on the clear button, in its context next to the filter input, implies that it's clearing the filter. I've had users who were very surprised when clicking it cleared their log history.

This commit updates the button to read `clear screen` instead, which should help with that confusion.